curve offset to async Task & YouTurn cleanup

+++ SourceCode/GPS/Forms/Guidance/FormQuickAB.cs
@@ -154,7 +154,7 @@ namespace AgOpenGPS
                     new vec2(mf.curve.desList[mf.curve.desList.Count - 1].easting,
                     mf.curve.desList[mf.curve.desList.Count - 1].northing);
 
-                mf.trk.gArr[idx].mode = (int)TrackMode.Curve;
+                mf.trk.gArr[idx].mode = TrackMode.Curve;
 
                 //calculate average heading of line
                 double x = 0, y = 0;
@@ -289,7 +289,7 @@ namespace AgOpenGPS
             mf.trk.gArr[idx].ptA = new vec2(mf.ABLine.desPtA);
             mf.trk.gArr[idx].ptB = new vec2(mf.ABLine.desPtB);
 
-            mf.trk.gArr[idx].mode = (int)TrackMode.AB;
+            mf.trk.gArr[idx].mode = TrackMode.AB;
 
             mf.trk.gArr[idx].heading = mf.ABLine.desHeading;
 
@@ -380,7 +380,7 @@ namespace AgOpenGPS
             mf.trk.gArr[idx].ptA = new vec2(mf.ABLine.desPtA);
             mf.trk.gArr[idx].ptB = new vec2(mf.ABLine.desPtB);
 
-            mf.trk.gArr[idx].mode = (int)TrackMode.AB;
+            mf.trk.gArr[idx].mode = TrackMode.AB;
 
             mf.trk.gArr[idx].heading = mf.ABLine.desHeading;
